+++ source/renderer/zettlr-preview.js
@@ -145,10 +145,11 @@ class ZettlrPreview {
 
       let sort = (d.type === 'directory' || d.type === 'virtual-directory') ? `data-sorting="${d.sorting}" ` : ''
       let selected = (this._selectedFile && this._selectedFile === d.hash) ? ` selected` : ''
+      let id = (d.id) ? `data-id=${d.id}` : ''
       let snippets = (this._snippets) ? ' snippets' : ''
       let vdclass = (inVirtualDir) ? ' vd-file' : '' // File is not actually present in this "dir"
       let vdhashAttr = (inVirtualDir && vdhash) ? ' data-vd-hash="' + vdhash + '"' : '' // For context menu actions we need to pass vd-hash, b/c accessing parent will not work.
-      let elem = `<li class="${d.type}${selected}${snippets}${vdclass}" data-hash="${d.hash}" ${sort}${bgcolor}${vdhashAttr}>`
+      let elem = `<li class="${d.type}${selected}${snippets}${vdclass}" ${id} data-hash="${d.hash}" ${sort}${bgcolor}${vdhashAttr}>`
       if (d.type === 'directory' || d.type === 'virtual-directory') {
         // Render a directory
         elem += d.name