+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Quicklook windows can now be "popped out" so that they are no longer bound to the main window but can be dragged onto different displays, etc.
 - Windows and Linux windows now follow macOS in having no native window frame, but instead they employ the same strategy as macOS: The toolbar is the top element inside the main window of Zettlr, featuring window controls and, additionally, a button to open the application menu from the toolbar.
 - Zettlr correctly selects words containing apostrophs so that you can correct them adequately without the app "forgetting" the l' or 'll-part (or similar) of the word.
+- There's now an option to copy a file's ID to clipboard, if the file has one.
 
 ## Under the hood